Добавил карту, префабы, экспа в флоте

+++ Assets/Scripts/Scenes/GameScene/Contexts/PlayerContext/Abstracts/Interfaces/IExperienced.cs
@@ -2,6 +2,6 @@ namespace PlayerContext.Abstract.Interfaces
 {
     internal interface IExperienced
     {
-        public void AddExperience(int value);
+        public void AddExperience(float value);
     }
 }
